+++ tests/s/test_schkbk.c
@@ -231,7 +231,7 @@ static void test_dgebak(void** state)
     }
 
     print_message("SGEBAK: %d cases, max residual = %.3e (case %d)\n",
-                  knt, rmax, lmax_resid);
+                  knt, (double)rmax, lmax_resid);
     if (ninfo > 0)
         print_message("  INFO errors: %d (case %d)\n", ninfo, lmax_info);
