+++ tests/s/test_sdrvpo.c
@@ -333,7 +333,7 @@ static void run_ddrvpo_single(int n, int imat, int iuplo, int ifact, int iequed)
         for (int k = 0; k < nt; k++) {
             if (result[k] >= THRESH) {
                 fail_msg("SPOSV UPLO=%s test %d failed: result=%e >= thresh=%e",
-                         uplo, k + 1, result[k], THRESH);
+                         uplo, k + 1, (double)result[k], (double)THRESH);
             }
         }
     }
@@ -398,18 +398,18 @@ static void run_ddrvpo_single(int n, int imat, int iuplo, int ifact, int iequed)
         for (int k = k1 - 1; k < NTESTS; k++) {
             if (result[k] >= THRESH) {
                 fail_msg("SPOSVX FACT=%s UPLO=%s EQUED=%c test %d: result=%e >= thresh=%e",
-                         fact, uplo, equed_inout, k + 1, result[k], THRESH);
+                         fact, uplo, equed_inout, k + 1, (double)result[k], (double)THRESH);
             }
         }
     } else {
         /* TRFCON case: only check tests 1, 6 */
         if (!prefac && result[0] >= THRESH) {
             fail_msg("SPOSVX FACT=%s UPLO=%s EQUED=%c test 1: result=%e >= thresh=%e",
-                     fact, uplo, equed_inout, result[0], THRESH);
+                     fact, uplo, equed_inout, (double)result[0], (double)THRESH);
         }
         if (result[5] >= THRESH) {
             fail_msg("SPOSVX FACT=%s UPLO=%s EQUED=%c test 6: result=%e >= thresh=%e",
-                     fact, uplo, equed_inout, result[5], THRESH);
+                     fact, uplo, equed_inout, (double)result[5], (double)THRESH);
         }
     }
 }
